requirements.txt
matplotlib==3.8.3
numpy==2.2.4
Pillow==11.1.0
streamlit==1.42.2
tensorflow==2.18.0
tensorflow_hub==0.16.1
